refactor: Successfully format whole configuration with alejandra
Motivation: Linter warnings about unused arguments and alejandra failing
due to an empty `common.nix` file. I removed the file and the
unused arguments and it now passes the formatter checks.
If `common.nix` was doing something that I'm missing, it needs at
least the minimal valid expression, the following is enough:

```nix
{}
```

I have only seen the naming of arguments while in the
`flake.nix` file, I removed the `@ args` from the
`common/default.nix` as it was unused anyways.